% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pgd.R
\name{plot.pgd}
\alias{plot.pgd}
\title{Surface plot of the \code{pgd} class}
\usage{
\method{plot}{pgd}(o, nmodes = dim(o$f$x)[2], phi = 45, theta = 45)
}
\arguments{
\item{o}{The \code{pgd} class.}

\item{nmodes}{The number of modes to be plotted.}

\item{phi}{Angle defining the viewing direction: azimuthal direction.}

\item{theta}{Angle defining the viewing direction: colatitude.}
}
\value{
Surface plot of the approximated result.
}
\description{
Surface plot of the \code{pgd} class
}
