% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_2D.R
\name{poisson_2D}
\alias{poisson_2D}
\title{PGD for the 2D Poisson's equation in Cartesian coordinates}
\usage{
poisson_2D(
  src,
  n,
  bc,
  mlim,
  tol = 0.001,
  maxiter = list(f_loop = 500, r_loop = 251)
)
}
\arguments{
\item{src}{List of two elements, \code{x} and \code{y}, each containing an inline function with only one input parameter. Each function returns a list with the separate variables of the source term.}

\item{n}{List of two elements, \code{x} and \code{y}, each containing the number of nodes of each separate variable.}

\item{bc}{List of two elements, \code{x} and \code{y}, each containing a vector with the boundaries of each separate variable (referred to the nodes).}

\item{mlim}{List of two elements, \code{x} and \code{y}, each containing a vector with the limits of the mesh of each separate variable.}

\item{tol}{Tolerance of the errors of for the F and R loops.}

\item{maxiter}{List of two elements, \code{f_loop} and \code{r_loop}, each containing the maximum number of loop iterations.}
}
\value{
List of \code{pgd} class with three elements: \code{f}, \code{alpha}, \code{coor}. \code{f} contains the modes of each separate variable. \code{alpha} contains the alpha values of each mode (already included in \code{f}). \code{coor} contains the separate Cartesian coordinates of each node.
}
\description{
PGD for the 2D Poisson's equation in Cartesian coordinates
}
